+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ![](https://raw.githubusercontent.com/JaxsonWang/WP-Editor.md/docs/screenshots/Interface-logo.jpg) 
 
+### Version：5.0.4
+
+ * 添加Prism Tomorrow Night主题风格
+ * 更换编辑器语法高亮引擎为Prism
+ * 修复debugger警告提示问题
+ * 添加KaTeX兼容选项
+
 ### Version：5.0.3
 
  * 修复单独加载KaTeX失效的问题